Restructure dashboard format
