+++ frontend/src/App.js
@@ -4,7 +4,6 @@ import HomePage from "./views/HomePage/HomePage";
 import { Box } from "@mui/material";
 import DashboardPage from "views/DashboardPage/DashboardPage";
 import SigninPage from "views/SigninPage/SigninPage";
-import MainPage from "views/DashboardPage/MainPage";
 
 function App() {
   return (
@@ -16,7 +15,6 @@ function App() {
             <Route path="/signup/*" element={<SignupPage />} />
             <Route path="/dashboard/*" element={<DashboardPage />} />
             <Route path="/signin/*" element={<SigninPage />} />
-            <Route path="/main/*" element={<MainPage />} />
           </Routes>
         </Router>
       </Box>
