+++ frontend/src/views/DashboardPage/MainComponent.js
@@ -1,0 +1,16 @@
+import * as React from "react";
+import "./mainComponent.scss";
+
+function MainComponent() {
+  return (
+    <div className="main">
+      <div className="flexbox-container">
+        <div className="flexbox-item">This is Placeholder content 1</div>
+        <div className="flexbox-item">This is Placeholder content 2</div>
+        <div className="flexbox-item">This is Placeholder content 3</div>
+      </div>
+    </div>
+  );
+}
+
+export default MainComponent;
