+++ frontend/src/views/DashboardPage/dashboardComponent.scss
