+++ frontend/src/views/DashboardPage/dashboardPage.scss
@@ -1,0 +1,34 @@
+.dashboard {
+    .dashboardTopbar {
+        display: flex;
+        justify-content: space-between;
+        flex-wrap: nowrap;
+        border-bottom: .5px solid lightgray;
+        padding-top: 5px;
+        background-color: white;
+
+        .item:not(:nth-child(2)) {
+            flex-basis: 0%;
+            flex-grow: 1;
+        }
+
+        .item:last-child {
+            text-align: right;
+        }
+
+        .tabs {
+            margin-left: 5px;
+        }
+
+        .user {
+            margin-top: 10px;
+            margin-right: 20px;
+        }
+
+        .logo {
+            img {
+                height: 44px;
+            }
+        }
+    }
+}
