+++ frontend/src/views/DashboardPage/mainComponent.scss
